requirements.txt
Flask==1.0.2
Flask-WTF==0.14.2
Jinja2==2.10
newspaper3k==0.2.8 
nltk==3.4
Flask-Cors==3.0.7 
Flask-Bcrypt==0.7.1
python-dotenv==0.10.1
validators==0.12.4